+++ example/example/settings.py
@@ -143,11 +143,14 @@ MFA_QUICKLOGIN=True        # Allow quick login for returning users by provide on
 MFA_HIDE_DISABLE=('',)     # Can the user disable his key (Added in 1.2.0).
 MFA_REDIRECT_AFTER_REGISTRATION="registered"
 MFA_SUCCESS_REGISTRATION_MSG="Go to Home"
+MFA_ALWAYS_GO_TO_LAST_METHOD = True
+MFA_ENFORCE_RECOVERY_METHOD = True
+MFA_RENAME_METHODS = {"RECOVERY":"Backup Codes","FIDO2":"Biometric Authentication"}
 PASSWORD_HASHERS = DEFAULT_PASSWORD_HASHERS #Comment if PASSWORD_HASHER already set
 PASSWORD_HASHERS += ['mfa.recovery.Hash']
 RECOVERY_ITERATION = 1 #Number of iteration for recovery code, higher is more secure, but uses more resources for generation and check...
 TOKEN_ISSUER_NAME="PROJECT_NAME"      #TOTP Issuer name
 
-U2F_APPID="https://localhost"    #URL For U2F
+U2F_APPID="https://localhost:9000"    #URL For U2F
 FIDO_SERVER_ID="localhost"      # Server rp id for FIDO2, it the full domain of your project
 FIDO_SERVER_NAME="TestApp"